+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed PHP Mess detector issues.
+
+### Changed
+- Code refactored for order service methods.
+
 ## [2.9.0] - 2021-07-30
 ### Added
 - Added automatic cancellation of MultiSafepay pretransactions for non-paid canceled orders.